move gpu_allocator's error impl out of dx12 to a shared location for vulkan+dx12
